feat(model): optimally implement a uniform real distribution (#24)
Due to the change from int to double, the old random function didn't
work.
Now the controller uses the newly-implemented Uniform struct to generate
uniformly generated doubles

+++ src/model/object/entity/individual/individual.cpp
@@ -8,4 +8,4 @@ Individual::Individual(double x, double y, double energy) :
 
 long Individual::getID() {return ID;}
 
-int Individual::getVisionRange() { return visionRange;}
+double Individual::getVisionRange() { return visionRange;}
